Disable feature-aware selection of devices for core features
This is because a bug in `get_physical_device_features_2` is appearing in a `next` chain when composed of core feature-related structures, and could not be tracked down.
This might be a bug in the wrapper, or in the NVIDIA drivers used for testing; will require a more extensive investigation, and time.